Add designers.yml, remove example files
