--- site/src/data/heroes.yml
@@ -1,3 +1,0 @@
-- name: Richard Stallman
-- name: Phil Katz
-- name: Ada Lovelace
